+++ software/foxglove_extensions/zyarm-control-panels/src/index.ts
@@ -1,12 +1,12 @@
 import { ExtensionContext } from "@foxglove/extension";
 
-import { initClawControlPanel } from "./ClawControlPanel";
+import { initGripperControlPanel } from "./GripperControlPanel";
 import { initMotorTemperaturePanel } from "./MotorTemperaturePanel";
 
 export function activate(extensionContext: ExtensionContext): void {
   extensionContext.registerPanel({
-    name: "ZYArm Claw Control",
-    initPanel: initClawControlPanel,
+    name: "ZYArm Gripper Control",
+    initPanel: initGripperControlPanel,
   });
   extensionContext.registerPanel({
     name: "ZYArm Motor Temperatures",